perf(posts): split summary client state
